Kestrel (RTL8852B/RTL8852C, Wi-Fi 6): one HAL for both dies on the verbatim-vendored halbb/halrf plane (#293)
## Kestrel (RTL8852BU / RTL8832CU) — one design for both dies, vendored
BB/RF plane

Refactors the Kestrel HAL so both chips share one architecture, with
Realtek's C reused verbatim wherever the vendor ships it:

- **mac_ax plane** (power-on / FWDL / efuse / MAC TRX / H2C-C2H / USB /
descriptors): hand-ported C++, both dies through the same code with
per-variant register banks (`_V1`) and descriptors.
- **BB/RF plane: Realtek halbb/halrf compiled VERBATIM** for both dies —
one chip-agnostic core + per-chip backends under `hal/halbb/g6/` +
`hal/halrf/g6/` (`vendor/` untouched vendor files, `shim/` +
`hal_headers_le.h` the minimal PHL surface, `kestrel_hal{bb,rf}_glue.c`
the C entry points, `tools/vendor_hal{bb,rf}_*.sh` re-vendor). The
vendored loaders own the BB/RF/gain tables, per-channel BB config, RF
tune, DACK/RX-DCK (+ IQK on the 8852C); radio pages reach the fw through
the bridge `send_h2c` into the mac_ax H2C encoder.
- All hand-rolled BB/RF fallbacks and the extract-generated table copies
are deleted (−54K lines): the vendored hwimg arrays are the only tables.
- Builds and CI-passes on **all platforms** — Linux GCC/Clang, macOS,
MinGW, **MSVC** — plus new `kestrel-8852b-only` / `kestrel-8852c-only`
CI cells.
- No milestones/WIP language; no usbmon-capture debris (the port is
source-matched to `reference/rtl8852bu` @ v1.19.21 and
`reference/rtl8852cu`).

### Real bugs found and fixed on the way

- **Cold-boot power-on** (both dies): a USB chip from real cold reads
`WLMAC_PWR_STE=MAC_ON` (its AFSM auto-powers to enumerate); the vendor
`pwr.c` prologue forces the MAC off first — without it the WCPU bootrom
never raises `H2C_PATH_RDY`. The auto-probing in-tree rtw89 had
pre-initialized the chip in every earlier session, masking this.
- **`HALBB_*_SUPPORT=0` defines**: the vendor tests them with `#ifdef`,
so `=0` compiled fw-offload IN and the 8852B BBCR writer applied zero
registers. Off = undefined.
- **Coex GNT from cold**: `B_AX_ENBT` must be set before any LTE-space
access; the GNT arbitration is now SW-forced to WiFi
(`mac_cfg_gnt_8852b` stance).
- **RFK prologue position is load-bearing**:
NCTL/SI-reset/LCK/RCK/efuse-trim now run at `phy_bb_rf_init` (the
vendor's `halrf_dm_init` position). Run lazily after the tune, the a-die
SI reset deafens the radio both ways (bisected on-air).
- **Shutdown SIGABRT**: `Stop()` now joins the WP-release drain thread
before the demos call `libusb_exit` (libusb `usbi_mutex_destroy`
assertion, 5/5 reproducible → 0/10).

### Evidence-gated calibration stance

TSSI/DPK (both dies) and per-channel IQK (8852B) are gated off: under
the fixed-dBm power model they measurably degrade TX (8852B IQK: paired
cold A/B on a 5 GHz MCS0 flood, 8812AU monitor — 10/1 decodes with IQK
vs 100/37 without). The 8852C validated on-air WITH IQK and keeps it.
Gates fall with a TSSI-referenced power model.

### Hardware validation (final ladder, VBUS-cold, rtw89
temp-blacklisted)

| Check | Result | Baseline |
|---|---|---|
| identity (kestrelprobe) | PASS both dies | — |
| 8852BU rx_smoke ch36 | 17 events | 17 |
| 8852BU tx_sdr duty ch6 / ch36 | 79.5% / 66.9% | 80.6% / 66.0% |
| narrowband occ-BW 20/10/5 | 15.6 / 8.2 / 4.1 MHz | 16.4 / 8.2 / 4.2 |
| bw_rx 20/40/80 + clean DACK | PASS | PASS |
| beacon_onair (HW TBTT) | PASS (11 txhits) | PASS |
| 8832CU rx_smoke ch36+ch6 | PASS | PASS |
| 8832CU TX matrix (6 cells incl. HE) | 6/6, 58–88% duty | 6/6, 79–88% |
| 5G-160 HE-MCS7 | 49.2% duty ≈ 66.4 Mbps | ~66.6 Mbps |
| 6G-80 HE-MCS7 | 45.3% duty | 45% |
| 6G-160 (documented limitation) | 2.1% duty (witness) | 0.8% |
| EVM MCS0 (8812AU monitor) | 8832CU −14 dB n=1606; 8852BU −6 dB n=769 |
— (absolute) |
| inject_sanity (rtw89 witness) | PASS | PASS |
| streamtx_sdr (video path) | 32.1% duty, no stall | PASS |
| init→first-RX-frame | 1.3 s (B) / 3.1 s (C) | no regression gate hit |
| regress.py ch6 + ch36 | 4/4 both | 4/4 |

Known limitation kept as documented: **6 GHz TX tops out at 80 MHz** on
the 8832CU (the 6G+160 TX-enable path is un-ported; RF locks and RX-160
works — B210-confirmed). MCS7 TX→monitor decode is bench-limited at the
current geometry (link EVM ≈ −14 dB at MCS0; both dies, pre- and
post-refactor identically).


---------

+++ src/RxPacket.h
@@ -42,7 +42,11 @@ struct rx_pkt_attrib
      * the chip's TSF against its own wall clock. Populated by
      * FrameParser; surfaced through examples/rx/main.cpp's rx.frame event. */
     uint32_t tsfl;
-    uint8_t data_rate;
+    /* AX/VHT datarate code is 9-bit (HT 0x80+, VHT 0x100+, HE 0x180+), so a
+     * uint8_t truncates the format bits — VHT2SS+ and every HE rate collapse
+     * onto an HT/legacy code. Kept 16-bit so the reported RX rate is truthful
+     * (needed to identify HE frames when a Kestrel acts as an HE monitor). */
+    uint16_t data_rate;
     uint8_t bw;
     uint8_t stbc;
     uint8_t ldpc;
